requirements.txt
anthropic
google-cloud-bigquery
yfinance
python-dotenv
